Relocate snapshot folder
