+++ components/common/PageLayout/components/Header/snapshot/InputVotingStrategies.tsx
